requirements.txt
# Kubernetes Python Client
kubernetes>=28.1.0

# YAML 처리
PyYAML>=6.0

# 로깅 및 유틸리티
python-dateutil>=2.8.2

# 타입 힌트 지원
typing-extensions>=4.5.0

# 개발 도구 (선택사항)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0